+++ pt.h
@@ -3,8 +3,8 @@
 
 struct pt
 	{
-	double x;
-	double y;
+	float x;
+	float y;
 	bool operator==( const pt & rhs )const{return (x==rhs.x && y == rhs.y); }
 	bool operator!=( const pt & rhs )const{return !(*this==rhs); }
 	};
